Introduce ColorHelper
Move the function `interpolateInRgb` into the new class `ColorHelper` as it will later also be needed when the LFO graph is made scalable.
